requirements.txt
streamlit
python-dotenv
google-generativeai
openai
anthropic
PyPDF2
docx2txt
tenacity